svg/geometry/parsing/r-invalid.svg

<svg xmlns="http://www.w3.org/2000/svg" xmlns:h="http://www.w3.org/1999/xhtml" width="800" height="8000"><title>SVG Geometry Properties: parsing r with invalid values</title><metadata><h:link href="https://svgwg.org/svg2-draft/geometry.html#R" rel="help"/><h:meta content="r supports only the grammar '&lt;length-percentage&gt;'." name="assert"/></metadata><h:script src="/resources/testharness.js"/><h:script src="/resources/testharnessreport.js"/><h:script src="/css/support/parsing-testcommon.js"/><script><![CDATA[

test_invalid_value("r", "10");
test_invalid_value("r", "auto");
test_invalid_value("r", "10px 20px");
test_invalid_value("r", "-1px");
test_invalid_value("r", "-10%");

  ]]></script></svg>
