LayoutTests/imported/w3c/web-platform-tests/svg/geometry/parsing/r-invalid-expected.txt


FAIL e.style['r'] = "10" should not set the property value assert_equals: expected "" but got "10px"
PASS e.style['r'] = "auto" should not set the property value 
PASS e.style['r'] = "10px 20px" should not set the property value 
FAIL e.style['r'] = "-1px" should not set the property value assert_equals: expected "" but got "-1px"
FAIL e.style['r'] = "-10%" should not set the property value assert_equals: expected "" but got "-10%"
